Graphics converted to using float data types.
- Graphics don't need to be super precise, float usually does the
  trick. Should hopefully speed things up a bit.

+++ include/appcommon/graphics/data_set_2d.h
@@ -17,36 +17,36 @@ class DataSet2d {
 
   /// \brief Gets the maximum x value.
   /// \return The maximum x value.
-  virtual double MaxX() const = 0;
+  virtual float MaxX() const = 0;
 
   /// \brief Gets the maximum y value.
   /// \return The maximum y value.
-  virtual double MaxY() const = 0;
+  virtual float MaxY() const = 0;
 
   /// \brief Gets the minimum x value.
   /// \return The minimum x value.
-  virtual double MinX() const = 0;
+  virtual float MinX() const = 0;
 
   /// \brief Gets the minimum y value.
   /// \return The minimum y value.
-  virtual double MinY() const = 0;
+  virtual float MinY() const = 0;
 
  protected:
   /// \var x_max_
   ///   The maximum x value.
-  mutable double x_max_;
+  mutable float x_max_;
 
   /// \var x_min_
   ///   The minimum x value.
-  mutable double x_min_;
+  mutable float x_min_;
 
   /// \var y_max_
   ///   The maximum y value.
-  mutable double y_max_;
+  mutable float y_max_;
 
   /// \var y_min_
   ///   The minimum value.
-  mutable double y_min_;
+  mutable float y_min_;
 };
 
 #endif  // OTLS_APPCOMMON_GRAPHICS_DATASET2D_H_
